Refactored entity

+++ src/main/java/com/carsharing/model/Rental.java
@@ -1,8 +1,13 @@
 package com.carsharing.model;
 
 import java.util.Date;
-
-import jakarta.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
@@ -12,7 +17,7 @@ import lombok.EqualsAndHashCode;
 public class Rental {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private Date rentalDate;
     private Date returnDate;
     private Date actualReturnDate;
